+++ Entities/Models/Location.cs
@@ -9,9 +9,10 @@ namespace Entities.Models
     {
         public string SystemName { get; set; }
         public string DetailedLocation { get; set; }
-        public float X { get; set; }
-        public float Y { get; set; }
-        public float Z { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Z { get; set; }
 
+        
     }
 }
